+++ common-ui/src/main/java/org/simple/clinic/common/ui/theme/SimpleColors.kt
@@ -11,8 +11,10 @@ data class SimpleColors(
     val toolbarPrimary: Color = Color.Unspecified,
     val toolbarPrimaryVariant: Color = Color.Unspecified,
     val onToolbarPrimary: Color = Color.Unspecified,
+    val onToolbarPrimary72: Color = onToolbarPrimary.copy(alpha = 0.72f),
     val material: Colors = lightColors(),
-    val onSurface67: Color = material.onSurface.copy(alpha = 0.67f)
+    val onSurface67: Color = material.onSurface.copy(alpha = 0.67f),
+    val onSurface11: Color = material.onSurface.copy(alpha = 0.11f)
 )
 
 internal val LocalSimpleColors = staticCompositionLocalOf { SimpleColors() }