+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Handle window insets when displaying app content in edge-to-edge
 - Enable edge-to-edge support on all versions of Android
 - Handle nullable inputs when removing last chip in the `ChipInputAutoCompleteTextView`
+- Migrate `PatientSummaryScreen` toolbar to Jetpack Compose
+- Add Compose components for Overdue screen
+- Migrate medicines summary view to Jetpack Compose
+- Migrate codeclimate config to qlty.sh config
+- Migrate `OverdueScreen` patient list to Jetpack Compose
 
 ## 2025.05.20